carve out list of known users
This list of users is an not sooo important
feature but exists to provide ownership in
cmdline. The list helps to know when a param
is a username.

However, it has the annoying fact that the
internal and public versions of tapper-autoreport
diverge unneccessarily. Now it optionally reads
the list from a file if that exists.