autoreport: skip_all now *really* skips all
So far it only adapted the resulting TAP but
the remaining script was still executed.

However, the require_* functions are there to
check for, well, requirements, and if they are
not met, the script should really skip everything
after that.

That's now solved.

We run the main function to create a report which
already knows about the SKIP conditions and additionally
exit after that.

En passant add utility function ok_or_skipall which
does an ok() but interprets its success as requirement.