tapper-autoreport: fix is_running_in_kvm_guest()
is_running_in_kvm_guest() tries to load the cpuid kernel module.
If this is already loaded or built into the kernel, this will fail
and will report back as "not running in KVM guest".
Extend the check by checking for the device file existence beforehand
and only try to load the module if this fails.
While at it flatting the routine by using early bail-out instead of
staircase like if cascades (feel free to revert this part if needed)