tapper-autoreport: add get_xen_tool() and use it
add the final version(tm) of detecting the proper Xen tool. This will
prefer xl over xm, returning the path to the tool.
The result will be shell-0 is at least one tool was found and 1 if
neither was available. Use it like this in functions:
    XL=$(get_xen_tool) || return
or like this in a test-script:
    XL=$(get_xen_tool)
    require_ok $? "Xen control tool \"$(basename $XENTOOL)\" used"

Use the new function whereever we open-coded the detection before.