tapper-autoreport: add disable_cpb() function
disable_cpb tries hard to disable the Core Performance Boost setting.
It should work in most cases, though mostly you need to be root for
actually changing the setting.
If running virtualized, this will mostly fail: CPB is not propagated
to guest, so it reports it as not-boosted, but is probably activated
in the host. Take care for yourself.
Outputs one diagnostic line to stdout and returns either 0 (success)
or 1 (failure).