tapper-autoreport: add MSR access helpers (read only for now)
Allows test scripts to read MSRs. This uses the /dev/cpu/<n>/msr
kernel interface, which requires root access and (probably) the
module to be loaded. To check for this before actually failing
there is a function has_msr_access which returns the status.
The actual read_msr() function takes the MSR number in hex as the
first parameter and optionally the CPU number as the second one,
defaulting to CPU 0. It outputs the content of the MSR as 16
unprefixed lower-case hex digits on stdout and returns 0 on success,
1 on an invalid MSR or 2 for file access problems.
This uses perl to do the proper byte-aligned seeking and value
conversion (dd does not work here).