tapper-autoreport: add write_msr routine
Add a write_msr routine. This is based on the read_msr function.
It takes both the MSR address (1st arg) and the new value (2nd arg)
as hex values and writes them to the CPU given in the optional
3rd arg (0 if this is omitted).
WARNING: use this function with care, you can easily crash the
machine by writing bogus values in some registers.