tapper-autoreport: avoid namespace cluttering, fixes feature detection
Some helper functions like has_kernel_config or has_cpufeature use
unimaginative variable names like "feature", which may clash (and
they did!) with variables from other, high-level parts of a test
script. One could solve this with using subshells for the functions
(rounded brackets instead of curly ones), but this is probably bloat
and not necessary. This patch renames the variables using a "_ta"
prefix to avoid accidental clashes in the future.
One should review the rest of tapper-autoreport for more candidates.