tapper-utils: frontend to call autoreport util functions
That is a frontend script to have one frontend
tool that executes all parameters in context
of loaded --import-utils. This can be used
to access the utilities from other programs
easily e.g. like `tapper-utils get_vendor`
or 'tapper-utils has_cpufeature fpu'.

Depending on the utility function you
evaluate the result output or the exit code
or both.

You are responsible to only use it for the
meaningful utility functions.