acpi-cpufreq: example test script
Test for joined AMD and Intel P-state driver

There is an effort to include AMD support in the current Intel
P-state driver called acpi-cpufreq.

This test is to test the functionality of this joined driver.
The hardware P-state support (starting with GH) is removed from
the powernow-k8 driver, so this must not load anymore.
The test is designed to run on AMD and Intel hardware, as we have
to spot regressions in the current Intel functionality as well.