more heuristical port for URL printing
When sending to Tapper the script prints out
the URL of the report. This URL is based on
assumptions and assembled together. We now
use an explicit variable for the port or assume
that on 'localhost' the port is 3000, as used
when starting the web gui with development servers.

Yes, that's just a heuristic, and it still can fail.