bash-test-utils: introduce a file_upload()
This is not really a direct file upload but more
a mark-up of files to be uploaded iff te TAP report
is packed into an archive for later sending to
a TAP database, like Tapper.

In that case the files marked with upload_file() are
added to that package and are later uploaded.

That is essentially the functionality from tapper-autoreport
but with a de-coupling of execution time and upload time.