bash-test-utils: allow user-specific progam/package mapping for install_program()
install_program() installs a program but knows
about mapping to actual package names which might
be distro-specific, eg. g++ in Debian is just g++
but in ALinux the package is gcc-c++.

bash-test-utils has a default collection for most
common uses and we now extend this by providing
additional user-specific mapping that an actual
test program needs.