bash-test-utils: [RFK] cleanup stdout-vs.-file confusion
Now also early exits like on a SKIPALL use
the done_testing() function which decides
where output goes.

This solves the problem that some scripts with
require_* functions were not uploaded and their
output was just part of a logfile - depending on
the actual success that the require_* function
checked.